refactoring structure services
